refactor(commit): エディタ起動処理をget-editor.tsへ分離
- エディタコマンド生成ロジックをkonoka-editor.tsからget-editor.tsへ移動
- konokaEdit内でgetEditorを利用するよう修正
- コードの責務分離により可読性と保守性を向上
