+++ plugins/commit/src/konoka-editor.ts
@@ -1,8 +1,9 @@
 import { Command, FileSystem } from "@effect/platform";
 import type { CommandExecutor } from "@effect/platform/CommandExecutor";
 import type { PlatformError } from "@effect/platform/Error";
-import { Config, Data, Effect } from "effect";
+import { Data, Effect } from "effect";
 import { appendDiffToEditmsg, removeDiffFromEditmsg } from "./attach-diff";
+import { getEditor } from "./get-editor";
 
 export class EditorFailedError extends Data.TaggedError("EditorFailedError")<{
   readonly editor: Command.Command;
@@ -14,65 +15,6 @@ export class EditorFailedError extends Data.TaggedError("EditorFailedError")<{
 }
 
 /**
- * 環境変数`EDITOR`が未設定または空のときに使うデフォルトのエディタコマンドです。
- */
-export const defaultEditor = "emacsclient --reuse-frame --alternate-editor=emacs" as const;
-
-/**
- * 環境変数`EDITOR`の値を取得し、
- * 未設定または空文字列の場合は`defaultEditor`を返します。
- *
- * `EDITOR`はユーザが設定するものなので信頼できる文字列として扱います。
- * シェルが解釈するコマンド文字列としてそのまま`sh -c`に渡される前提です。
- */
-export const editorCommand: Effect.Effect<string> = Config.nonEmptyString("EDITOR").pipe(
-  Effect.orElseSucceed(() => defaultEditor),
-);
-
-/**
- * `EDITOR`コマンド文字列とファイルパスから、
- * `sh -c`経由で起動するための`argv`を生成します。
- *
- * gitやsudoeditと同じく、
- * `EDITOR`はシェルが解釈するコマンド文字列として扱います。
- * 実行ファイルパスにスペースがある場合はユーザがクオートで対処します。
- * ファイルパスは`"$@"`経由の位置引数として渡し、
- * シェル側での再解釈を避けます。
- */
-export function buildEditorInvocation(
-  editor: string,
-  file: string,
-): readonly ["sh", "-c", string, "konoka-editor", string] {
-  return [
-    "sh",
-    "-c",
-    `${editor} "$@"`,
-    // `sh -c <script> <argv0> <argv1>...`の引数列で、
-    // `<argv0>`はスクリプト内の`$0`に入ります。
-    // シェルがエラーメッセージを出す際の名乗りに使われるので、ここを`konoka-editor`にしておくと、
-    // `konoka-editor: <editor>: not found`のように出所の追跡しやすい表示になります。
-    "konoka-editor",
-    file,
-  ];
-}
-
-/**
- * テキストエディタを起動するためのコマンドを構築します。
- */
-function newEditor(commitEditmsgPath: string): Effect.Effect<Command.Command, PlatformError> {
-  return Effect.gen(function* () {
-    const editor = yield* editorCommand;
-    const [executable, ...args] = buildEditorInvocation(editor, commitEditmsgPath);
-    const cmd = Command.make(executable, ...args).pipe(
-      Command.stdin("inherit"),
-      Command.stdout("inherit"),
-      Command.stderr("inherit"),
-    );
-    return cmd;
-  });
-}
-
-/**
  * コマンドを推定して起動して編集を行います。
  */
 export function konokaEdit(
@@ -80,7 +22,7 @@ export function konokaEdit(
   patchPath: string,
 ): Effect.Effect<void, PlatformError | EditorFailedError, CommandExecutor | FileSystem.FileSystem> {
   return Effect.gen(function* () {
-    const editor = yield* newEditor(commitEditmsgPath);
+    const editor = yield* getEditor(commitEditmsgPath);
     yield* appendDiffToEditmsg(commitEditmsgPath, patchPath);
     const code = yield* Command.exitCode(editor); // テキストエディタが実際に起動します。
     yield* removeDiffFromEditmsg(commitEditmsgPath);
